fix: corrige bugs do dashboard /backoffice
- Receita sempre $0: filtro de pagamentos usava status 'completed'
  (minusculo) em vez de PAGAMENTO_STATUS.COMPLETED, nunca batendo com
  nenhum registro real
- Campo revenue era Int mas o valor somado tem centavos; mudado para
  Float (mesmo padrao usado em Payment.amount/Plan.price)
- Distribuicao de planos contava o campo legado User.membership
  (sempre 'FREE' no seed) em vez da assinatura real do usuario
- Distribuicao de roles: usuarios com role desconhecida/nula eram
  silenciosamente contados como BARBERSHOP_OWNER; extraida logica
  de mapeamento duplicada para um helper unico com warning de log
- Grafico semanal de crescimento retornava dados fake zerados;
  agora calcula contagem real dos ultimos 7 dias
- allCompletedPayments: removidas 3 queries de debug descartaveis e
  todos os filtros comentados ("testar"); filtros de plano, usuario,
  data de pagamento, proximo pagamento, status e metodo restaurados
  como filtros reais do Prisma
- Removidos console.log de debug (incluindo um dentro de loop por
  usuario) em favor do SmartLogger
- Removido BackofficeController, um controller REST vazio sem uso
  desde a migracao para GraphQL

+++ src/graphql/types/backoffice.type.ts
@@ -1,4 +1,4 @@
-import { ObjectType, Field, Int, registerEnumType } from '@nestjs/graphql';
+import { ObjectType, Field, Int, Float, registerEnumType } from '@nestjs/graphql';
 import { UserRole, MembershipStatus } from './enums';
 
 export enum UserStatus {
@@ -22,7 +22,7 @@ export class BackofficeStats {
   @Field(() => Int)
   newUsersThisMonth: number;
 
-  @Field(() => Int)
+  @Field(() => Float)
   revenue: number;
 }
